IB/uverbs: add object ref count to uobject
the ref count describe the usage of the 'object' pointer of this
uobject with in the current ib context. each time another uobject
with in the current ib context need this uobject, the ref count
is bumped. unlike the ib_xxx usecnt which is not directly related
to the current ib context, this value is only valid within single ib
context.

the uobj abstraction layer will use this value to protect the uobjects
from too early destruction within the current ib context.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>