IB/uverbs: shared uobjects must not destroy the ib_x object
shared uobjects are not the owners of the HW ib_x object and
thus they are not allowed to free this object. this commit
prepare the uverbs for object sharing and make sure that even
with object sharing, the ib_x destruction will not be affected.

when shared object is deleted, the uobj is deleted and the ib_x
HW object ref count is decremented.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>