IB/core: Add context use count for ib_x HW objects
This commit add the initial code to support share of ib_x HW objects
across multiple ib_ucontext. When created, the ib_x HW object belongs
only to the ib_ucontext that created it. At this point all the locking
added to uverbs and ib core protect this object from race with destroy
(e.g disassociation) is in place and no change is needed.

Once the ib_x HW object is shared with another ib_ucontext, each context
now see this ib_x HW object as local to the context without caring/knowing
if this ib_x HW object is shared or not. Attempt to destroy the object from
the new context might cause pre-mature destruction of the ib_x HW object
for the other context.

To mitigate this, additional refcount need be added to each ib_x HW object
that allow ib core to decide if it can destroy the ib_x HW object. This
imformation is exposed to the ib core via the ib_uobject.

When ib core skip the destroy of ib_x HW object because it is still used by
another ib_ucontext, it will not leak the HW object resources because the
other context will eventually destroy this object but this time the destroy
will not be skipped.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>