IB/uverbs: Add ib_XX export to file provider
This provider allow exporting of ib core HW object (ib_XX)
from one user context to another. the exported object has new
uobject in the new context which point to the same HW object.
Both uobjects point to the ib_xx object sharecnt counter. When
ib_xx object is shared, this counter is incremented. When uobject
is destroyed, uverbs_destroy_uobject check if the uobject has valid
sharecnt pointer. If there is no pointer, the uobject is assumed to
be not shared and the rest of the object destruction can be performed
using the per context locks. However, if sharecnt pointer is valid,
the function atomic_dec_and_test the counter and only if this is
the last ref of the ib_xx object, it call type_class->destroy_hw
on the uobject which in turn invoke the ib_xx destroy function. If
the shrecnt is valid and > 0, the call to type_class->destroy_hw is
skipped and only the uverbs resources are destroyed.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>