RDMA/uverbs: uobj_put_obj_read macro should be removed
uobj_put_obj_read macro assume that ib_x has pointer to the
ib_uobject. this is wrong assumption as future patches will
remove the ib_uobject pointers from the ib_x objects. now
that we fixed the uobj_get_obj_read macro and we have the
ib_uobject we no longer need to use the ib_x to find the
ib_uobject and we can remove the uobj_put_obj_read macro.

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>