IB/verbs: Define prototype for clone callback
Define prototype for clone callback. The clone callback will be
part of any shareable IB HW object (ib_pd, ib_mr,...) and will be
used by the driver layer to supply the uverbs a way to clone IB HW
object driver data to rdma-core user space provider. The clone
callback will be used when new IB HW object is created and every
time it is imported to some ib_ucontext.

Rules for the clone callback are:

1. IB HW objects that do not have the clone callback are not
shareable across all the drivers

2. IB HW objects that do have clone callback are shareable only
for drivers that set this pointer to some valid callback
function

Signed-off-by: Shamir Rabinovitch <shamir.rabinovitch@oracle.com>